FEAT: Add functionality for product ProductDetails
Getting single product details when click on a product.
