+++ package.json
@@ -23,7 +23,7 @@
     "next": "14.2.3",
     "react": "^18",
     "react-dom": "^18",
-    "react-hook-form": "^7.51.4",
+    "react-hook-form": "^7.51.5",
     "react-icons": "^5.2.1",
     "react-redux": "^9.1.2",
     "react-slick": "^0.30.2",
