+++ src/app/products/[id]/page.tsx
@@ -1,4 +1,5 @@
-import type { Metadata, ResolvingMetadata } from "next";
+import ProductDetails from "@/components/custom/ProductDetails";
+import type { Metadata } from "next";
 type Props = {
   params: { id: string };
   searchParams: { [key: string]: string | string[] | undefined };
@@ -12,8 +13,21 @@ export async function generateMetadata({ params }: Props): Promise<Metadata> {
   };
 }
 
-const productDetails = () => {
-  return <section className="container">productDetails</section>;
+const getSingleData = async (id: any) => {
+  const response = await fetch(
+    `https://rezauls-json-server.vercel.app/clothings/${id}`
+  );
+  return await response.json();
+};
+
+const productDetails = async ({ params }: any) => {
+  const product = await getSingleData(params.id);
+  console.log(product);
+  return (
+    <section className="container">
+      <ProductDetails product={product} />
+    </section>
+  );
 };
 
 export default productDetails;
