+++ src/components/custom/ProductDetails.tsx
@@ -1,4 +1,9 @@
-const ProductDetails = () => {
+import { Product } from "@/lib/productType";
+
+interface ProductProp {
+  product: Product;
+}
+const ProductDetails = ({ product }: ProductProp) => {
   return <div>ProductDetails</div>;
 };
