+++ src/components/custom/TopSelling.tsx
@@ -4,29 +4,7 @@ import Link from "next/link";
 import { Button } from "../ui/button";
 import { useAppSelector } from "@/redux/hooks/hooks";
 import { useEffect, useState } from "react";
-
-type Review = {
-  username: string;
-  comment: string;
-  rating: number;
-};
-interface Product {
-  id: number;
-  title: string;
-  brand: string;
-  description: string;
-  size: string[];
-  color: string[];
-  price: number;
-  thumbnail: string;
-  images: string[];
-  category: string;
-  style: string;
-  search: string[];
-  stock: number;
-  ratings: number;
-  reviews: Review[];
-}
+import { Product } from "@/lib/productType";
 
 const TopSelling = () => {
   const { products } = useAppSelector((state) => state.products);
